Split buyData into different Jobs
