+++ buyer-api/src/facades/helpers/index.js
@@ -5,4 +5,5 @@ export {
   performTransaction,
   sendTransaction,
   getTransactionReceipt,
+  retryAfterError,
 } from './performTransaction';
